demos/voice_notes/requirements.txt
transformers
accelerate
torchaudio